Implement `storage`, `defaultDomain`. Refactoring.
