+++ Sources/Mailgun/Extensions/Mailgun+Encode.swift
@@ -1,0 +1,12 @@
+extension Mailgun {
+    func encode(apiKey: String) throws -> String {
+        guard let apiKeyData = "api:\(apiKey)".data(using: .utf8) else {
+            throw MailgunError.encodingProblem
+        }
+        let authKey = apiKeyData.base64EncodedData()
+        guard let authKeyEncoded = String.init(data: authKey, encoding: .utf8) else {
+            throw MailgunError.encodingProblem
+        }
+        return authKeyEncoded
+    }
+}
